adding filtering and pagination

+++ CarPlates/DTOs/ExecutedPlateDto.cs
@@ -5,12 +5,12 @@ namespace CarPlates.DTOs
 {
 	public class ExecutedPlateDto
 	{
-		public int Id { get; set; } = 0;
+		public long Id { get; set; } = 0;
 		public string? Letters { get; set; }
 		public string? Numbers { get; set; }
 		public int ExecutionYear { get; set; }
 		public int ExecutionNumber { get; set; }
 		public DateTime Date { get; set; }
-		public int CarTypeId { get; set; }
+		public long CarTypeId { get; set; }
 	}
 }
